Added score to the game. Over time the score gradually increases. This increment increases in value the longer the player has been alive (based off player speed).

+++ Classes/ScoreManager.cpp
@@ -18,17 +18,17 @@ ScoreManager* ScoreManager::sharedScoreManager()
 	return instance;
 }
 
-void ScoreManager::addToScore(int point)
+void ScoreManager::addToScore(float point)
 {
 	score += point;
 }
 
-void ScoreManager::subtractFromScore(int point)
+void ScoreManager::subtractFromScore(float point)
 {
 	score -= point;
 }
 
-int ScoreManager::getScore()
+float ScoreManager::getScore()
 {
 	return score;
 }
